feat: move commands out of the server namespace
BREAKING CHANGE: target, container-registry and provider commands are no longer part of the server namespace
