+++ pkg/cmd/apikey/generate.go
