+++ pkg/cmd/containerregistry/containerregistry.go
