+++ pkg/cmd/containerregistry/delete.go
