+++ pkg/cmd/containerregistry/set.go
