+++ pkg/cmd/provider/uninstall.go
