+++ pkg/server/providers.go
@@ -78,7 +78,7 @@ func (s *Server) registerProviders() error {
 			}
 
 			if manager.HasUpdateAvailable(info.Name, info.Version, *manifest) {
-				log.Infof("Update available for %s. Update with `daytona server provider update`.", info.Name)
+				log.Infof("Update available for %s. Update with `daytona provider update`.", info.Name)
 			}
 		}
 	}
