+++ pkg/views/server/config.go
@@ -46,7 +46,7 @@ func RenderConfig(config *server.Config) {
 	output += "If you want to connect to the server remotely:\n\n"
 
 	output += "1. Create an API key on this machine: "
-	output += lipgloss.NewStyle().Foreground(views.Green).Render("daytona server api-key new") + "\n"
+	output += lipgloss.NewStyle().Foreground(views.Green).Render("daytona api-key new") + "\n"
 	output += "2. Add a profile on the client machine: \n\t"
 	output += lipgloss.NewStyle().Foreground(views.Green).Render(fmt.Sprintf("daytona profile add -a %s -k API_KEY", apiUrl))
