+++ pkg/views/util/help.go
@@ -39,7 +39,7 @@ func getLongDescriptionText() string {
 		"  Use the following commands to get started:\n\n" +
 		fmt.Sprintf("  \x1b[1m%-*s\x1b[0m%s", helpDescriptionLabelWidth, "1) daytona server", "Start the Daytona Server process locally\n") +
 		fmt.Sprintf("  \x1b[1m%-*s\x1b[0m%s", helpDescriptionLabelWidth, "2) daytona git-providers add", "Register a Git provider of your choice\n") +
-		fmt.Sprintf("  \x1b[1m%-*s\x1b[0m%s", helpDescriptionLabelWidth, "3) daytona server target set", "Set a target to spin up your Dev Environments on\n") +
+		fmt.Sprintf("  \x1b[1m%-*s\x1b[0m%s", helpDescriptionLabelWidth, "3) daytona target set", "Set a target to spin up your Dev Environments on\n") +
 		fmt.Sprintf("  \x1b[1m%-*s\x1b[0m%s", helpDescriptionLabelWidth, "4) daytona ide", "Choose the default IDE\n") +
 		fmt.Sprintf("  \x1b[1m%-*s\x1b[0m%s", helpDescriptionLabelWidth, "5) daytona whoami", "Show information about the currently logged in user\n") +
 		fmt.Sprintf("  \n%s\x1b[1m%s\x1b[0m\n\n", "That's it! Start coding - ", "daytona create")
@@ -58,7 +58,7 @@ func getLongDescriptionFull() string {
 			fmt.Sprintf("%s\n", "     @@@@ @@  @@@@@@@@@@@   ") +
 			fmt.Sprintf("%s\x1b[1m%-*s\x1b[0m%s", " @@@@@@@@        @@@        ", helpDescriptionLabelWidthWithSigil, "1) daytona server", "Start the Daytona Server process locally\n") +
 			fmt.Sprintf("%s\x1b[1m%-*s\x1b[0m%s", "  @@@@@          @@@@@      ", helpDescriptionLabelWidthWithSigil, "2) daytona git-providers add", "Register a Git provider of your choice\n") +
-			fmt.Sprintf("%s\x1b[1m%-*s\x1b[0m%s", "    @@@@@       @@@@@@@@    ", helpDescriptionLabelWidthWithSigil, "3) daytona server target set", "Set a target to spin up your Dev Environments\n") +
+			fmt.Sprintf("%s\x1b[1m%-*s\x1b[0m%s", "    @@@@@       @@@@@@@@    ", helpDescriptionLabelWidthWithSigil, "3) daytona target set", "Set a target to spin up your Dev Environments\n") +
 			fmt.Sprintf("%s\x1b[1m%-*s\x1b[0m%s", "     @@@        @@@@@@@@@   ", helpDescriptionLabelWidthWithSigil, "4) daytona ide", "Choose the default IDE\n") +
 			fmt.Sprintf("%s\x1b[1m%-*s\x1b[0m%s", "  @@@@@@@@@ @@@@@@@@  @@    ", helpDescriptionLabelWidthWithSigil, "5) daytona whoami", "Show information about the currently logged in user\n") +
 			fmt.Sprintf("%s\n", "  @@@@@@@@@@@@@@@@@@        ") +
